+++ src/main/java/com/revivalmodding/revivalcore/core/client/models/ModelSuit.java
@@ -2,235 +2,173 @@ package com.revivalmodding.revivalcore.core.client.models;
 
 import net.minecraft.client.model.ModelBase;
 import net.minecraft.client.model.ModelBiped;
+import net.minecraft.client.model.ModelPlayer;
 import net.minecraft.client.model.ModelRenderer;
-import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
-import net.minecraft.inventory.EntityEquipmentSlot;
-import net.minecraft.util.EnumHand;
-import net.minecraft.util.EnumHandSide;
+import net.minecraft.entity.item.EntityArmorStand;
+import net.minecraft.entity.monster.EntityZombie;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.MathHelper;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 
-public class ModelSuit extends ModelBase {
+public class ModelSuit extends ModelPlayer {
 
-    protected final EntityEquipmentSlot slot;
-    public boolean isSneak;
-    public ModelBiped.ArmPose leftArmPose;
-    public ModelBiped.ArmPose rightArmPose;
-    protected ModelRenderer bipedHead;
-    protected ModelRenderer bipedBody;
-    protected ModelRenderer bipedRightArm;
-    protected ModelRenderer bipedLeftArm;
-    protected ModelRenderer bipedRightLeg;
-    protected ModelRenderer bipedLeftLeg;
-    private boolean modelPrepared;
+	public boolean smallArms;
+	private static ModelBiped cachedModel;
 
-    public ModelSuit(EntityEquipmentSlot slot) {
-        ModelRenderer EMPTY = new ModelRenderer(this);
-        this.slot = slot;
-        this.modelPrepared = false;
+	public ModelRenderer bipedLeftArmwear;
+	public ModelRenderer bipedRightArmwear;
+	public ModelRenderer bipedLeftLegwear;
+	public ModelRenderer bipedRightLegwear;
+	public ModelRenderer bipedBodyWear;
 
-        this.bipedHead = EMPTY;
-        this.bipedBody = EMPTY;
-        this.bipedRightArm = EMPTY;
-        this.bipedLeftArm = EMPTY;
-        this.bipedRightLeg = EMPTY;
-        this.bipedLeftLeg = EMPTY;
+	public ModelSuit(float f, boolean hasSmallArms) {
+		super(f, hasSmallArms);
+
+		this.textureWidth = 64;
+		this.textureHeight = 64;
+		
+		float scale = f + 0.5F;
+
+        if (hasSmallArms) {
+            this.bipedRightArm = new ModelRenderer(this, 40, 16);
+            this.bipedRightArm.addBox(-2.0F, -2.0F, -2.0F, 3, 12, 4, f);
+            this.bipedRightArm.setRotationPoint(-5.0F, 2.5F, 0.0F);
+
+            this.bipedLeftArm = new ModelRenderer(this, 32, 48);
+            this.bipedLeftArm.addBox(-1.0F, -2.0F, -2.0F, 3, 12, 4, f);
+            this.bipedLeftArm.setRotationPoint(5.0F, 2.5F, 0.0F);
+
+            this.bipedLeftArmwear = new ModelRenderer(this, 48, 48);
+            this.bipedLeftArmwear.addBox(-1.0F, -2.0F, -2.0F, 3, 12, 4, 0.25F);
+            this.bipedLeftArmwear.setRotationPoint(0.0F, 0.0F, 0.0F);
+
+            this.bipedRightArmwear = new ModelRenderer(this, 40, 32);
+            this.bipedRightArmwear.addBox(-2.0F, -2.0F, -2.0F, 3, 12, 4, 0.25F);
+            this.bipedRightArmwear.setRotationPoint(0.0F, 0.0F, 0.0F);
+        } else {
+            this.bipedLeftArm = new ModelRenderer(this, 32, 48);
+            this.bipedLeftArm.addBox(-1.0F, -2.0F, -2.0F, 4, 12, 4, f);
+            this.bipedLeftArm.setRotationPoint(5.0F, 2.0F, 0.0F);
+
+            this.bipedLeftArmwear = new ModelRenderer(this, 48, 48);
+            this.bipedLeftArmwear.addBox(-1.0F, -2.0F, -2.0F, 4, 12, 4, 0.25F);
+            this.bipedLeftArmwear.setRotationPoint(0.0F, 0.0F, 0.0F);
+
+            this.bipedRightArmwear = new ModelRenderer(this, 40, 32);
+            this.bipedRightArmwear.addBox(-3.0F, -2.0F, -2.0F, 4, 12, 4, 0.25F);
+            this.bipedRightArmwear.setRotationPoint(0.0F, 0.0F, 0.0F);
+        }
+
+        this.bipedLeftLeg = new ModelRenderer(this, 16, 48);
+        this.bipedLeftLeg.addBox(-2.0F, 0.0F, -2.0F, 4, 12, 4, f);
+        this.bipedLeftLeg.setRotationPoint(1.9F, 12.0F, 0.0F);
+
+        this.bipedLeftLegwear = new ModelRenderer(this, 0, 48);
+        this.bipedLeftLegwear.addBox(-2.0F, 0.0F, -2.0F, 4, 12, 4, 0.25F);
+        this.bipedLeftLegwear.setRotationPoint(0.0F, 0.0F, 0.0F);
+
+        this.bipedRightLegwear = new ModelRenderer(this, 0, 32);
+        this.bipedRightLegwear.addBox(-2.0F, 0.0F, -2.0F, 4, 12, 4, 0.25F);
+        this.bipedRightLegwear.setRotationPoint(0.0F, 0.0F, 0.0F);
+
+        this.bipedBodyWear = new ModelRenderer(this, 16, 32);
+        this.bipedBodyWear.addBox(-4.0F, 0.0F, -2.0F, 8, 12, 4, 0.25F);
+        this.bipedBodyWear.setRotationPoint(0.0F, 0.0F, 0.0F);
+
+        this.bipedBody.addChild(bipedBodyWear);
+        this.bipedRightArm.addChild(bipedRightArmwear);
+        this.bipedLeftArm.addChild(bipedLeftArmwear);
+        this.bipedRightLeg.addChild(bipedRightLegwear);
+        this.bipedLeftLeg.addChild(bipedLeftLegwear);
     }
 
     @Override
-    public void render(Entity entityIn, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch, float scale) {
-        this.setRotationAngles(limbSwing, limbSwingAmount, ageInTicks, netHeadYaw, headPitch, scale, entityIn);
-        GlStateManager.pushMatrix();
-
-        if (this.isChild) {
-            float f = 2.0F;
-            GlStateManager.scale(0.75F, 0.75F, 0.75F);
-            GlStateManager.translate(0.0F, 16.0F * scale, 0.0F);
-            this.bipedHead.render(scale);
-            GlStateManager.popMatrix();
-            GlStateManager.pushMatrix();
-            GlStateManager.scale(0.5F, 0.5F, 0.5F);
-            GlStateManager.translate(0.0F, 24.0F * scale, 0.0F);
-            this.bipedBody.render(scale);
-            this.bipedRightArm.render(scale);
-            this.bipedLeftArm.render(scale);
-            this.bipedRightLeg.render(scale);
-            this.bipedLeftLeg.render(scale);
-        } else {
-            if (entityIn.isSneaking()) {
-                GlStateManager.translate(0.0F, 0.2F, 0.0F);
-            }
-
-            this.bipedHead.render(scale);
-            this.bipedBody.render(scale);
-            this.bipedRightArm.render(scale);
-            this.bipedLeftArm.render(scale);
-            this.bipedRightLeg.render(scale);
-            this.bipedLeftLeg.render(scale);
-        }
-
-        GlStateManager.popMatrix();
+    public void render(Entity entity, float f, float f1, float f2, float f3, float f4, float f5)
+    {
+        super.render(entity, f, f1, f2, f3, f4, f5);
     }
 
+    @Override
     public void setRotationAngles(float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch, float scaleFactor, Entity entityIn) {
-        boolean flag = entityIn instanceof EntityLivingBase && ((EntityLivingBase) entityIn).getTicksElytraFlying() > 4;
-        this.bipedHead.rotateAngleY = netHeadYaw * 0.017453292F;
+        super.setRotationAngles(limbSwing, limbSwingAmount, ageInTicks, netHeadYaw, headPitch, scaleFactor, entityIn);
 
-        if (flag) {
-            this.bipedHead.rotateAngleX = -((float) Math.PI / 4F);
-        } else {
-            this.bipedHead.rotateAngleX = headPitch * 0.017453292F;
+        if (entityIn instanceof EntityArmorStand) {
+            EntityArmorStand entityarmorstand = (EntityArmorStand) entityIn;
+            this.bipedHead.rotateAngleX = 0.019F * entityarmorstand.getHeadRotation().getX();
+            this.bipedHead.rotateAngleY = 0.019F * entityarmorstand.getHeadRotation().getY();
+            this.bipedHead.rotateAngleZ = 0.019F * entityarmorstand.getHeadRotation().getZ();
+            this.bipedHead.setRotationPoint(0.0F, 1.0F, 0.0F);
+            this.bipedBody.rotateAngleX = 0.019F * entityarmorstand.getBodyRotation().getX();
+            this.bipedBody.rotateAngleY = 0.019F * entityarmorstand.getBodyRotation().getY();
+            this.bipedBody.rotateAngleZ = 0.019F * entityarmorstand.getBodyRotation().getZ();
+            this.bipedLeftArm.rotateAngleX = 0.019F * entityarmorstand.getLeftArmRotation().getX();
+            this.bipedLeftArm.rotateAngleY = 0.019F * entityarmorstand.getLeftArmRotation().getY();
+            this.bipedLeftArm.rotateAngleZ = 0.019F * entityarmorstand.getLeftArmRotation().getZ();
+            this.bipedRightArm.rotateAngleX = 0.019F * entityarmorstand.getRightArmRotation().getX();
+            this.bipedRightArm.rotateAngleY = 0.019F * entityarmorstand.getRightArmRotation().getY();
+            this.bipedRightArm.rotateAngleZ = 0.019F * entityarmorstand.getRightArmRotation().getZ();
+            this.bipedLeftLeg.rotateAngleX = 0.019F * entityarmorstand.getLeftLegRotation().getX();
+            this.bipedLeftLeg.rotateAngleY = 0.019F * entityarmorstand.getLeftLegRotation().getY();
+            this.bipedLeftLeg.rotateAngleZ = 0.019F * entityarmorstand.getLeftLegRotation().getZ();
+            this.bipedLeftLeg.setRotationPoint(1.9F, 11.0F, 0.0F);
+            this.bipedRightLeg.rotateAngleX = 0.019F * entityarmorstand.getRightLegRotation().getX();
+            this.bipedRightLeg.rotateAngleY = 0.019F * entityarmorstand.getRightLegRotation().getY();
+            this.bipedRightLeg.rotateAngleZ = 0.019F * entityarmorstand.getRightLegRotation().getZ();
+            this.bipedRightLeg.setRotationPoint(-1.9F, 11.0F, 0.0F);
+            copyModelAngles(this.bipedHead, this.bipedHeadwear);
         }
 
-        this.bipedBody.rotateAngleY = 0.0F;
-        this.bipedRightArm.rotationPointZ = 0.0F;
-        this.bipedRightArm.rotationPointX = -5.0F;
-        this.bipedLeftArm.rotationPointZ = 0.0F;
-        this.bipedLeftArm.rotationPointX = 5.0F;
-        float f = 1.0F;
-
-        if (flag) {
-            f = (float) (entityIn.motionX * entityIn.motionX + entityIn.motionY * entityIn.motionY + entityIn.motionZ * entityIn.motionZ);
-            f = f / 0.2F;
-            f = f * f * f;
+        if(entityIn instanceof EntityZombie) {
+            boolean flag = entityIn instanceof EntityZombie && ((EntityZombie)entityIn).isArmsRaised();
+            float f = MathHelper.sin(this.swingProgress * (float)Math.PI);
+            float f1 = MathHelper.sin((1.0F - (1.0F - this.swingProgress) * (1.0F - this.swingProgress)) * (float)Math.PI);
+            this.bipedRightArm.rotateAngleZ = 0.0F;
+            this.bipedLeftArm.rotateAngleZ = 0.0F;
+            this.bipedRightArm.rotateAngleY = -(0.1F - f * 0.6F);
+            this.bipedLeftArm.rotateAngleY = 0.1F - f * 0.6F;
+            float f2 = -(float)Math.PI / (flag ? 1.5F : 2.25F);
+            this.bipedRightArm.rotateAngleX = f2;
+            this.bipedLeftArm.rotateAngleX = f2;
+            this.bipedRightArm.rotateAngleX += f * 1.2F - f1 * 0.4F;
+            this.bipedLeftArm.rotateAngleX += f * 1.2F - f1 * 0.4F;
+            this.bipedRightArm.rotateAngleZ += MathHelper.cos(ageInTicks * 0.09F) * 0.05F + 0.05F;
+            this.bipedLeftArm.rotateAngleZ -= MathHelper.cos(ageInTicks * 0.09F) * 0.05F + 0.05F;
+            this.bipedRightArm.rotateAngleX += MathHelper.sin(ageInTicks * 0.067F) * 0.05F;
+            this.bipedLeftArm.rotateAngleX -= MathHelper.sin(ageInTicks * 0.067F) * 0.05F;
         }
 
-        if (f < 1.0F) {
-            f = 1.0F;
-        }
+        if (cachedModel != null) {
+            copyModelAngles(cachedModel.bipedHead, this.bipedHead);
+            copyModelAngles(cachedModel.bipedBody, this.bipedBody);
+            copyModelAngles(cachedModel.bipedRightArm, this.bipedRightArm);
+            copyModelAngles(cachedModel.bipedLeftArm, this.bipedLeftArm);
+            copyModelAngles(cachedModel.bipedRightLeg, this.bipedRightLeg);
+            copyModelAngles(cachedModel.bipedLeftLeg, this.bipedLeftLeg);
 
-        this.bipedRightArm.rotateAngleX = MathHelper.cos(limbSwing * 0.6662F + (float) Math.PI) * 2.0F * limbSwingAmount * 0.5F / f;
-        this.bipedLeftArm.rotateAngleX = MathHelper.cos(limbSwing * 0.6662F) * 2.0F * limbSwingAmount * 0.5F / f;
-        this.bipedRightArm.rotateAngleZ = 0.0F;
-        this.bipedLeftArm.rotateAngleZ = 0.0F;
-        this.bipedRightLeg.rotateAngleX = MathHelper.cos(limbSwing * 0.6662F) * 1.4F * limbSwingAmount / f;
-        this.bipedLeftLeg.rotateAngleX = MathHelper.cos(limbSwing * 0.6662F + (float) Math.PI) * 1.4F * limbSwingAmount / f;
-        this.bipedRightLeg.rotateAngleY = 0.0F;
-        this.bipedLeftLeg.rotateAngleY = 0.0F;
-        this.bipedRightLeg.rotateAngleZ = 0.0F;
-        this.bipedLeftLeg.rotateAngleZ = 0.0F;
-
-        if (this.isRiding) {
-            this.bipedRightArm.rotateAngleX += -((float) Math.PI / 5F);
-            this.bipedLeftArm.rotateAngleX += -((float) Math.PI / 5F);
-            this.bipedRightLeg.rotateAngleX = -1.4137167F;
-            this.bipedRightLeg.rotateAngleY = ((float) Math.PI / 10F);
-            this.bipedRightLeg.rotateAngleZ = 0.07853982F;
-            this.bipedLeftLeg.rotateAngleX = -1.4137167F;
-            this.bipedLeftLeg.rotateAngleY = -((float) Math.PI / 10F);
-            this.bipedLeftLeg.rotateAngleZ = -0.07853982F;
-        }
-
-        this.bipedRightArm.rotateAngleY = 0.0F;
-        this.bipedRightArm.rotateAngleZ = 0.0F;
-
-        switch (this.leftArmPose) {
-            case EMPTY:
-                this.bipedLeftArm.rotateAngleY = 0.0F;
-                break;
-            case BLOCK:
-                this.bipedLeftArm.rotateAngleX = this.bipedLeftArm.rotateAngleX * 0.5F - 0.9424779F;
-                this.bipedLeftArm.rotateAngleY = 0.5235988F;
-                break;
-            case ITEM:
-                this.bipedLeftArm.rotateAngleX = this.bipedLeftArm.rotateAngleX * 0.5F - ((float) Math.PI / 10F);
-                this.bipedLeftArm.rotateAngleY = 0.0F;
-        }
-
-        switch (this.rightArmPose) {
-            case EMPTY:
-                this.bipedRightArm.rotateAngleY = 0.0F;
-                break;
-            case BLOCK:
-                this.bipedRightArm.rotateAngleX = this.bipedRightArm.rotateAngleX * 0.5F - 0.9424779F;
-                this.bipedRightArm.rotateAngleY = -0.5235988F;
-                break;
-            case ITEM:
-                this.bipedRightArm.rotateAngleX = this.bipedRightArm.rotateAngleX * 0.5F - ((float) Math.PI / 10F);
-                this.bipedRightArm.rotateAngleY = 0.0F;
-        }
-
-        if (this.swingProgress > 0.0F) {
-            EnumHandSide enumhandside = this.getMainHand(entityIn);
-            ModelRenderer modelrenderer = this.getArmForSide(enumhandside);
-            float f1 = this.swingProgress;
-            this.bipedBody.rotateAngleY = MathHelper.sin(MathHelper.sqrt(f1) * ((float) Math.PI * 2F)) * 0.2F;
-
-            if (enumhandside == EnumHandSide.LEFT) {
-                this.bipedBody.rotateAngleY *= -1.0F;
-            }
-
-            this.bipedRightArm.rotationPointZ = MathHelper.sin(this.bipedBody.rotateAngleY) * 5.0F;
-            this.bipedRightArm.rotationPointX = -MathHelper.cos(this.bipedBody.rotateAngleY) * 5.0F;
-            this.bipedLeftArm.rotationPointZ = -MathHelper.sin(this.bipedBody.rotateAngleY) * 5.0F;
-            this.bipedLeftArm.rotationPointX = MathHelper.cos(this.bipedBody.rotateAngleY) * 5.0F;
-            this.bipedRightArm.rotateAngleY += this.bipedBody.rotateAngleY;
-            this.bipedLeftArm.rotateAngleY += this.bipedBody.rotateAngleY;
-            this.bipedLeftArm.rotateAngleX += this.bipedBody.rotateAngleY;
-            f1 = 1.0F - this.swingProgress;
-            f1 = f1 * f1;
-            f1 = f1 * f1;
-            f1 = 1.0F - f1;
-            float f2 = MathHelper.sin(f1 * (float) Math.PI);
-            float f3 = MathHelper.sin(this.swingProgress * (float) Math.PI) * -(this.bipedHead.rotateAngleX - 0.7F) * 0.75F;
-            modelrenderer.rotateAngleX = (float) ((double) modelrenderer.rotateAngleX - ((double) f2 * 1.2D + (double) f3));
-            modelrenderer.rotateAngleY += this.bipedBody.rotateAngleY * 2.0F;
-            modelrenderer.rotateAngleZ += MathHelper.sin(this.swingProgress * (float) Math.PI) * -0.4F;
-        }
-
-        if (this.isSneak) {
-            this.bipedBody.rotateAngleX = 0.5F;
-            this.bipedRightArm.rotateAngleX += 0.4F;
-            this.bipedLeftArm.rotateAngleX += 0.4F;
-            this.bipedRightLeg.rotationPointZ = 4.0F;
-            this.bipedLeftLeg.rotationPointZ = 4.0F;
-            this.bipedRightLeg.rotationPointY = 9.0F;
-            this.bipedLeftLeg.rotationPointY = 9.0F;
-            this.bipedHead.rotationPointY = 1.0F;
-        } else {
-            this.bipedBody.rotateAngleX = 0.0F;
-            this.bipedRightLeg.rotationPointZ = 0.1F;
-            this.bipedLeftLeg.rotationPointZ = 0.1F;
-            this.bipedRightLeg.rotationPointY = 12.0F;
-            this.bipedLeftLeg.rotationPointY = 12.0F;
-            this.bipedHead.rotationPointY = 0.0F;
-        }
-
-        this.bipedRightArm.rotateAngleZ += MathHelper.cos(ageInTicks * 0.09F) * 0.05F + 0.05F;
-        this.bipedLeftArm.rotateAngleZ -= MathHelper.cos(ageInTicks * 0.09F) * 0.05F + 0.05F;
-        this.bipedRightArm.rotateAngleX += MathHelper.sin(ageInTicks * 0.067F) * 0.05F;
-        this.bipedLeftArm.rotateAngleX -= MathHelper.sin(ageInTicks * 0.067F) * 0.05F;
-
-        if (this.rightArmPose == ModelBiped.ArmPose.BOW_AND_ARROW) {
-            this.bipedRightArm.rotateAngleY = -0.1F + this.bipedHead.rotateAngleY;
-            this.bipedLeftArm.rotateAngleY = 0.1F + this.bipedHead.rotateAngleY + 0.4F;
-            this.bipedRightArm.rotateAngleX = -((float) Math.PI / 2F) + this.bipedHead.rotateAngleX;
-            this.bipedLeftArm.rotateAngleX = -((float) Math.PI / 2F) + this.bipedHead.rotateAngleX;
-        } else if (this.leftArmPose == ModelBiped.ArmPose.BOW_AND_ARROW) {
-            this.bipedRightArm.rotateAngleY = -0.1F + this.bipedHead.rotateAngleY - 0.4F;
-            this.bipedLeftArm.rotateAngleY = 0.1F + this.bipedHead.rotateAngleY;
-            this.bipedRightArm.rotateAngleX = -((float) Math.PI / 2F) + this.bipedHead.rotateAngleX;
-            this.bipedLeftArm.rotateAngleX = -((float) Math.PI / 2F) + this.bipedHead.rotateAngleX;
+            cachedModel = null;
         }
     }
 
-    protected ModelRenderer getArmForSide(EnumHandSide side) {
-        return side == EnumHandSide.LEFT ? this.bipedLeftArm : this.bipedRightArm;
-    }
+    @Override
+    public void setModelAttributes(ModelBase model) {
 
-    protected EnumHandSide getMainHand(Entity entityIn) {
-        if (entityIn instanceof EntityLivingBase) {
-            EntityLivingBase entitylivingbase = (EntityLivingBase) entityIn;
-            EnumHandSide enumhandside = entitylivingbase.getPrimaryHand();
-            return entitylivingbase.swingingHand == EnumHand.MAIN_HAND ? enumhandside : enumhandside.opposite();
-        } else {
-            return EnumHandSide.RIGHT;
+        this.textureWidth = 64;
+        this.textureHeight = 64;
+
+        super.setModelAttributes(model);
+
+        if (model instanceof ModelBiped) {
+            cachedModel = (ModelBiped) model;
         }
     }
 
-    // TODO
-    protected ModelBiped.ArmPose getArmPose(Entity entity) {
-        return ModelBiped.ArmPose.EMPTY;
+
+
+    public void setModelVisibility() {
+        this.setVisible(true);
     }
 }
